+++ kernel/src/drv/disk/dpm.c
@@ -79,7 +79,7 @@ void dpm_metadata_write(char Letter, uint32_t Addr){
  * 
  * @return Кол-во прочитанных байт
  */
-size_t dpm_read(char Letter, uint32_t high_offset, uint32_t low_offset, size_t Size, void *Buffer){
+size_t dpm_read(char Letter, uint64_t high_offset, uint64_t low_offset, size_t Size, void *Buffer){
 	int Index = Letter - 65;
 
 	Index = (Index > 32 ? Index - 32 : Index);
@@ -120,7 +120,7 @@ size_t dpm_read(char Letter, uint32_t high_offset, uint32_t low_offset, size_t S
  * 
  * @return size_t - Кол-во записанных байт
  */
-size_t dpm_write(char Letter, uint32_t high_offset, uint32_t low_offset, size_t Size, char* Buffer){
+size_t dpm_write(char Letter, uint64_t high_offset, uint64_t low_offset, size_t Size, char* Buffer){
 	int Index = Letter - 65;
 
 	Index = (Index > 32 ? Index - 32 : Index);
